fixup type compatibility checking for map, function types
(Fixes up commit 49c7df18f8630013a842bd2524a829e11defb3ff)

Say type T_1 is compatible with type T_2 if T_2 can be substituted
for T_1. Notate this T_2 <= T_1.

For function and map types, we had type compatibility defined
by something like

(A->B) <= (C->D) iff (A <= C) and (B <= D),

where (A->B) denotes the function type which takes A and yields B.

But it should be

(A->B) <= (C->D) iff (C <= A) and (B <= D).

This is because if a function f is usable as type (C->D) it needs
to be the case that a C can be used as an argument for the function,
and that the result of the function can be returned as a D.

This commit changes the behavior for function types and map types,
and adds some unit tests. I believe that it works correctly for
overloaded functions via inheritance but I didn't add unit tests
for this.