when typechecking fails, show more type information (WIP)
This commit adds a function to get the type and pinpoint location
for all subexpressions of an expression. It does this recursively
for all children of the failed expression. This kind of dump can
sometimes help to quickly spot the problem.

It might be better if this was the behavior when a "verbose" flag
is passed to anura or something. It also would be more useful if
it were used in more places.