Care more on error formatting on type errors.
variant.cpp: Providing 'generate_error_2(2)' almost same as
'generate_error(1)', but 'generate_error_2(2)' accepts two different
messages, one to be printed to console, and the other one to
be passed to 'SDL_ShowSimpleMessageBox(4)'. 'generate_error(1)' prints
the same string both to console and the SDL box, this leads to a badly
stacked SDL box if the message is very long and/or the screen is
small. 'generate_error_2(2)' can be used to get wide console output
while caring better on message stacking in an SDL box.
'SDL_ShowSimpleMessageBox(4)' handles badly large strings.

variant.cpp: 'generate_error_2(2)' does the same as
'generate_error(1)' but relying on a new 'ASSERT_LOG_3(3)' instead of
'ASSERT_LOG(2)'.

asserts.hpp, asserts.cpp: Providing 'ASSERT_LOG_3(3)' to be able to
log assertions with different strings for console and SDL.
'ASSERT_LOG_3(3)' does the same as 'ASSERT_LOG(2)' but passing down
the arguments to 'report_assert_msg_2(2)' instead of
'report_assert_msg(1)'.

asserts.hpp, asserts.cpp: Providing 'report_assert_msg_2(2)' that does
the same as 'report_assert_msg(1)' but instead of printing the same
message to console and the SDL simple box, receives two different
strings, the first to be printed to console and the second to be
passed to 'SDL_ShowSimpleMessageBox(4)'. This is because
'SDL_ShowSimpleMessageBox(4)' handles badly too wide messages, while
text logs and console output fits naturally and safely lines of
arbitrary length.