lua: refactor how formula callable pointer is placed in userdata
In this commit, we basically change the way that formula callables
are passed as userdata to lua to make it "more RAII". Instead of
allocating space for a raw pointer to a formula callable, and
explicitly calling intrusive_ptr add ref, we just use the usual
lua idiom for passing a C++ class as a "value type" to lua, except
that class is boost::intrusive_ptr<FormulaCallable>. That is, the
userdata is created with size exactly matching one of these, but
we use placement new to construct the intrusive ptr at the
userdata position. When the gc metamethod is called, we explicitly
call the destructor of the type. This is slightly more transparent
because we don't explicitly have to remember to call add ref /
remove ref on the pointer.