lua can bind to Formula Object, use defn to convert lua vals to ffl
Add LuaContext::SetSelfObject. This uses a custom metatable which
uses features specific to FormulaObjects to give better behavior
than we get with FormulaCallables. Since FormulaObjects have class
definitions, we know the expected types of their fields. We add
a metatable for FormulaObjects whose __newindex metamethod will
check what type we should get before we try to convert the lua
value -- if it is Map of Lists vs. List of Maps etc. , we will try
to interpret nested lua tables correctly according to the desired
type signature.

We also add a helper function to class FormulaObject which makes
it easy to determine the expected assignment type of a property.