i18n: fixup handling of "" in po parser
In gettext, msgstr "" is supposed to indicate "untranslated", not
"translated as empty string".

We fix this, and also add support so that msgstr "\0" indicates
"translated as empty string".

More generally, any strings from a po file which contain embedded
nulls are terminated, and the nulls don't make it into the catalog.