Fix a Linux compilation error
A comparison between boost::optional and nullptr in src/tiled/tmx_reader.cpp
causes compilation to fail with boost version 1.58. This fixes it by
comparing with boost::none instead, which works in both the latest boost
and the bundled version.

Although this fix is not necessary for the bundled boost in the
compilation stage, on a Linux system the linking may fail because the
boost library on the system doesn't match the bundled boost headers.