Makefile: Add SANITIZE_UNDEFINED and config info
Make it possible to disable -fsanitize=undefined, which is
required for e.g. linking against musl libc or for the aarch64
target. For these targets there is neither libasan nor libubsan.

Print the currently active configuration values before actually
building the object files. This makes it easier for the packager
to see if everything is as expected.