Added support for batched drawing.
By setting batch_id in an object you can batch draws of objects with that Id. Also a host of other optimizations such as moving zorder to entity so zorder() calls are non-virtual and can be inlined since that's a pretty big deal for sorting objects by zorder. Added some in-engine profiler improvements as well.