Legalize identifier 'd'.
I haven't still seen actual uses of a space between the number of
rolls and the dice 'd', ever less in the documentation, but I think
they are allowed both before and after this. If I am not mistaken,
spaces between the dice 'd' and the side count of the dice are not
allowed either before or after this.

Activates the 'd' as a valid identifier, closing the gap in the first
graphies of the alphabet ('a', 'b', 'c', 'e'...) as being all valid
FFL identifiers.

Found while doing FFL debugging. This source could not be executed
previously unless by renaming variable 'd'.

    type(d)
    where c = [0]
    where d = '0'
    where type =
      def(a) -> string
        if (a is int,
            'int',
            if (a is string,
                'string',
                'fixme'))

Variable names like d<DIGIT><ALNUM>* are not allowed either before or
after this commit.