Do not count references to decimal variants.
At branch 'dev-no-decimal-refcount-mgmt'.

Given decimal typed variants are immutable - at least they very
probably they should, and that is congruent with the inner code of the
functions releasing and increasing reference counts
('variant::release()' and 'variant::increment_refcount()',
respectively); do not use unnecesarily the reference count management
sources on them, for a symbolic performance increase and an increased
harmony of the sources.

Changes are somewhat fine grain and self evident, but even though,
this commit is specifically passing the Anura automated test suite,
and both Frogatto and Argentum Age seem to be mostly OK with these
changes.