Bring back the off line mode to the splash screen.
Bring back the off line mode to the splash screen, now only meant to
be used when the developer mode is enabled. It is not that anyone
looking for a game is going to interrupt developers, it is to provide
developers the option to abstract themselves without resorting to
cutting network access at an operating system level.