Align the server status to the game title.
Cosmetic change for the `citadel` module. It should better be done in
the `citadel` module, but since the `tictactoe` and `wesnoth2` (at
least the part of `wesnoth2` that is working at my host) modules are
neutral to this change, it should do no harm.

Does this break your game? File an issue at GitHub!