Unbreak tests at my local Steam Runtime (1).
This is starting to look like there is a bug in
`decimal::from_string(...)` when being passed a string representing
integer, but if so, only manifesting at the Steam Runtime.

In other words:

`decimal::from_string("1.0")` is working as expectable from any host.

`decimal::from_string("1")` at my macOS and Linux hosts is working as
I could expect was expecting it to do. But it is yielding weird
results in the Steam Runtime at The Argent Lark and in the Steam
Runtime at my Linux host.