Apply Soliton suggestion of using `printf`.
There are some `zsh` configurations on Fedora that react bad to the
previous use of `echo`. Maybe that can be handled by trying to use
`/bin/echo` instead of shell defaults, and/or detect the current
running shell, but give first a chance to `printf`, as Soliton
suggested.

Revert the reversion of imgui that 6c709c01b368e17973 (previous rev)
did, I don't know of the reason for that reversion, so it looks like
just a cautious choice when adding that rev, but the reverted commit
moving `imgui` from `beb0ea7b47a4620c` to `d7ce9212f1d321a1` is safe
and actually useful to some of the environments I use.