Don't use ASSERT_LOG directly in main()
SDL_main is marked noexcept, but ASSERT_LOG was throwing uncaught exceptions which caused compile warnings.
Instead of asserting, an error will now be logged and main() returned. A more robust solution would be to
wrap most of main() in a catch-all block.

This also changes auto_select_resolution to take references instead of pointers since the function doesn't
work if they're null anyway, so there's no need for those parameters to be pointers.