Spent 5 minutes and got rid of the need for master-config.cfg.
This is a convenience for anyone making a module of their own.
It will be ignored by Git if you add one now - you should be able to store
it in your module and copy or link it to the base folder. You can still
include it in a release, of course. :)

# Conflicts:
#	.gitignore