Globally define GLM_ENABLE_EXPERIMENTAL for dynamic linux builds (#411)
For whatever reason, our old code now needs to globally define
`GLM_ENABLE_EXPERIMENTAL` to work with a newer version of GLM in the
build system. This does not appear to break any older builds, but still
flagged it with an `XXX` as a potentially unoptimal thing to eventually
get rid of. In fact it oddly enough re-enabled some of our previously
broken builds.

Used this opportunity to also seek out and better document the tail end
limits of the CI system.

Builds enabled by this:

* Fedora 30
* Fedora 31
* Fedora 41
* Debian 10 / Buster

Buillds potentially enabled by this, but not in the CI:

* Ubuntu 18.04 and older
* The glibc shipped there is too old for node 20 and modern GHA, like
checkout, run on that
* Fedora 29 and older
* Fedora does not ship the `dnf` package manager per default in versions
older than 30
* Debian 9 and older
* The minimal Docker image points at a mirror out of the box, which has
been archived since
* RHEL 7 and older based distros, like CentOS / Scientific Linux /
Oracle Linux
  * RHEL quasi abandoned the old CentOS model
* We cover Rocky and Alma from 8 onwards instead of bothering with the
past

And for the sake of completeness: older SuSe Leap versions are broken
for reasons unrelated to GLM.

I will not commit to complicating the build system any further to allow
for continuously testing for the very long tail of distribution support
as one would start needing per era or per version tricks in the build
matrix and the correct thing for users on those versions is to switch to
a newer OS and move on.