Replace vcpkg with Conan in the Steam Linux Runtime builds (#445)
In practice vcpkg has proven too flaky and weird to manage and I'd
rather spend the energy on fixing Anura instead of debugging various
transient build efforts.

Additionally with Conan we are able to bump dependencies independently
instead of bumping the whole baseline forward as we have had to with
vcpkg.

Frustrations around finding a common baseline for Steam Linux Runtime,
macOS and Windows is what finally made me motivated enough to go for
this. One effectively has to be lucky around the dependencies randomly
lining up in a baseline across the whole set of dependencies, which
often leads to situations where they effectively remain perpetually
frozen in time. I want to do better.